+++ fastmsa/templates/app/app/routes/sample.py.j2
@@ -1,6 +1,6 @@
 from fastmsa.api import get
 
-from fastmsa.repository import SqlAlchemyRepository
+from fastmsa.repo import SqlAlchemyRepository
 from {{msa.name}}.domain.models import User
 
 @get('/users')
